requirements.txt
customtkinter==5.2.2
Pillow==10.0.1
requests==2.31.0
